Move max_tokens field rewrites into protocol transform ops layer
These functions are pure transformations on *openai.ChatCompletionNewParams —
they belong next to the other OpenAI request transforms (cursor_compat
normalization, DeepSeek / Gemini / Moonshot adapters), not in server/.

- New: internal/protocol/transform/ops/request_openai_max_tokens.go with
  ApplyMaxCompletionTokensRewrite and ApplyMaxTokensRewrite (exported)
- New: matching _test.go with all unit + wire serialization tests
- server/openai.go: call ops.ApplyMax... instead of the local helpers
- server/rule_flags.go: keep only resolveRuleFlags and shouldStripUsage
  (dispatch-layer helpers that read *typ.Rule and reqCtx.Extra)
- server/rule_flags_test.go: drop the moved test cases
